Move LightCurveEstimator
