+++ gammapy/estimators/__init__.py
@@ -7,3 +7,4 @@ from .asmooth_map import *
 from .sensitivity import *
 from .flux_point import *
 from .profile import *
+from .lightcurve import *
