+++ gammapy/estimators/lightcurve.py
