+++ gammapy/time/__init__.py
@@ -1,7 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Time analysis."""
 from .lightcurve import *
-from .lightcurve_estimator import *
 from .period import *
 from .simulate import *
 from .variability import *
